backend/requirements.txt
Django==4.2.16
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
gunicorn==23.0.0
whitenoise==6.7.0
